Changed importing min to importing full stylesheet.
Changed secondary color in variables.less to match new blue.

Because the portfolio section is on blue now, I changed the hover to white. That can always be changed again.

Added success class to portfolio div and removed it from about div.

Also changed the color of the stars to match the new backgrounds.